Arm: Create dedicated module for arm things
