+++ emu/src/cpu/mod.rs
@@ -1,12 +1,10 @@
-mod alu_instruction;
+mod arm;
 pub mod arm7tdmi;
 mod condition;
 mod cpu_modes;
 mod data_processing;
 mod flags;
-mod instruction;
 mod move_compare_add_sub;
-mod opcode;
 mod psr;
 mod register_bank;
 mod registers;
